+++ src/test/converter2/issues/gh2721.ts
@@ -1,0 +1,3 @@
+export const big = 123n;
+
+export const neg = -123n;